+++ oekg/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class SparqlQueryConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "sparql_query"
+    name = "oekg"
