+++ oekg/views.py
@@ -5,11 +5,11 @@ from django.shortcuts import render
 from django.views.decorators.http import require_GET
 
 from oeplatform.settings import OEKG_SPARQL_ENDPOINT_URL
-from sparql_query.utils import validate_sparql_query
+from oekg.utils import validate_sparql_query
 
 
 def main_view(request):
-    response = render(request, "sparql_query/main.html")
+    response = render(request, "oekg/main.html")
     response["Content-Type"] = "text/html; charset=utf-8"
     return response
